auth done

+++ src/tasks/tasks.service.ts
@@ -1,11 +1,11 @@
-import { Injectable, NotFoundException } from "@nestjs/common";
-import { InjectRepository } from "@nestjs/typeorm";
-// import { v1 as uuid } from 'uuid';
-import { CreateTaskDto } from "./DTOs/create.task.dto";
-import { GetTaskFilterDto } from "./DTOs/get-task-filter.dto";
-import { TaskStatus } from "./task-status.enum";
-import { Task } from "./task.entity";
-import { TaskRepository } from "./task.repository";
+import { Injectable, NotFoundException } from '@nestjs/common';
+import { InjectRepository } from '@nestjs/typeorm';
+import { User } from 'src/auth/user.entity';
+import { CreateTaskDto } from './DTOs/create.task.dto';
+import { GetTaskFilterDto } from './DTOs/get-task-filter.dto';
+import { TaskStatus } from './task-status.enum';
+import { Task } from './task.entity';
+import { TaskRepository } from './task.repository';
 
 @Injectable()
 export class TasksService {
@@ -14,32 +14,38 @@ export class TasksService {
     private taskRepository: TaskRepository,
   ) {}
 
-  async getTasks(filterDto: GetTaskFilterDto): Promise<Task[]> {
-    return this.taskRepository.getTasks(filterDto);
+  async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
+    return this.taskRepository.getTasks(filterDto, user);
   }
 
-  async getTaskByID(id: number): Promise<Task> {
-    const found = await this.taskRepository.findOne(id);
+  async getTaskByID(id: number, user: User): Promise<Task> {
+    const found = await this.taskRepository.findOne({
+      where: { id, userId: user.id },
+    });
     if (!found) {
       throw new NotFoundException(`Task with ${id} not found`);
     }
     return found;
   }
 
-  async CreateTask(createTaskDto: CreateTaskDto): Promise<Task> {
-    return this.taskRepository.CreateTask(createTaskDto);
+  async CreateTask(createTaskDto: CreateTaskDto, user: User): Promise<Task> {
+    return this.taskRepository.CreateTask(createTaskDto, user);
   }
 
-  async DeteTaskById(id: number): Promise<void> {
-    const result = await this.taskRepository.delete(id);
+  async DeteTaskById(id: number, user: User): Promise<void> {
+    const result = await this.taskRepository.delete({ id, userId: user.id });
 
     if (result.affected === 0) {
       throw new NotFoundException(`Task with ID ${id} not found`);
     }
   }
 
-  async UpdateTaskStatus(id: number, status: TaskStatus): Promise<Task> {
-    const task = await this.getTaskByID(id);
+  async UpdateTaskStatus(
+    id: number,
+    status: TaskStatus,
+    user: User,
+  ): Promise<Task> {
+    const task = await this.getTaskByID(id, user);
     task.status = status;
     await task.save();
     return task;
